LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/security/embedded-credentials.tentative.sub-expected.txt



FAIL Embedded credentials are treated as network errors. assert_unreached: 'onload' should not fire. Reached unreachable code
FAIL Embedded credentials are treated as network errors in frames. assert_equals: The image should not load. expected "Error" but got "Load"
FAIL Embedded credentials are treated as network errors in new windows. assert_equals: The image should not load. expected "Error" but got "Load"
PASS Embedded credentials matching the top-level are not treated as network errors for relative URLs.
PASS Embedded credentials matching the top-level are not treated as network errors for same-origin URLs.
FAIL Embedded credentials matching the top-level are treated as network errors for cross-origin URLs. assert_equals: The image should load. expected "Error" but got "Load"
